backend/requirements.txt
fastapi==0.115.11
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
passlib==1.7.4
pydantic==2.9.2
bcrypt==4.1.3  # Старая версия для совместимости
pyjwt==2.9.0